+++ app/javascript/components/Article.jsx
@@ -4,6 +4,11 @@ import ReactDOM from 'react-dom'
 
 
 class Article extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { lu: this.props.Lu };
+  }
+
   render () {
     console.log("péage 4", this.props.article)
     return (
