+++ app/javascript/components/Fluxes.jsx
@@ -3,6 +3,7 @@ import PropTypes from "prop-types"
 import ReactDOM from 'react-dom'
 import axios from 'axios'
 import Article from './Article'
+import Flux from './Flux'
 
 function DisplayFlux(props) {
   console.log("péage1", props)
@@ -13,7 +14,7 @@ function DisplayFlux(props) {
             <h2>{flux.Title}</h2>
           </div>
           <ul>
-            <DisplayArticles id={flux.id}/>
+            <Flux id={flux.id}/>
           </ul>
         </div>)
       }
@@ -22,42 +23,6 @@ function DisplayFlux(props) {
     <div className="row margin-top" id="fluxes">
       {listflux}
     </div>
-  );
-}
-function DefineArticles(props) {
-  const url = "/articles?id=" + props.id
-  const request = axios.get(url)
-  const articles = request.then((response)=>{
-    response.data.map((article)=>{
-      return (<ul className="card-overflow" key={article.id}>
-        <Article article={article}/>
-        </ul>
-      );
-    })
-    return (
-      {articles}
-    )
-  })
-  return (
-    <ul className="card-overflow" key={article.id}>
-      {articles}
-    </ul>
-  );
-};
-
-function DisplayArticles(props) {
-  const articles = DefineArticles(props)
-  console.log("péage3", articles)
-  const listarticles = articles.map((article)=>{
-    return (<ul className="card-overflow" key={article.id}>
-      <Article article={article}/>
-      </ul>
-    );
-  });
-  return (
-    <ul className="card-overflow" key={article.id}>
-      {listarticles}
-    </ul>
   );
 }
